backend/requirements.txt
Django>=4.0,<5.0
djangorestframework>=3.13,<3.14
djangorestframework-simplejwt>=5.2,<6.0
channels>=3.0,<4.0
channels-redis>=3.3,<3.4
psycopg2-binary>=2.9,<3.0
django-environ>=0.8,<0.9
django-cors-headers>=3.13,<4.0
requests>=2.28,<3.0
gunicorn>=20.1,<21.0
flake8>=4.0,<5.0
